refactor(booking): use OffsetDateTime instead of LocalDateTime in Booking
Decided to use OffsetDateTime so there's a match between api.yml and the entity model
